Disable DDS support in aliceHLTWrapper, Fixes #183
DDS Intercom API has been changed in version v1.4 without maintaining
backward compatibility. As it is not a matter of changing a few lines
of code, DDS support is disabled now in the aliceHLTWrapper. The device
needs a major rewrite to adopt to the latest changes in FairRoot and
the new DDS Intercom API.