Track interface adapted to use std::array
Track objects added to the dictionaries, CATracker adapted to the new
interfaces.
The GetParam() and GetCov() methods are now commented: when and if there
is the necessity of them hopefully we shall see what is the best
implementation.
The explicit copy constructor of TrackPar has been removed: the compiler
is able to implement it correctly and without that definition it is free
to define also the operator= and the move constructor.