Adding static length check to the String2<Type> function, fixes #240
The String2<type> function takes a string and converts it to an integer
number of the specified type at compile time. static assert has been added
to check that the number of charcters in the string is not larger than the
size of the type.

Added also static_assert with reasonable error message if the type is
a non-integral type.

Deleted obsolete functions for specific data types.