Basic configuration for simulation / An o2sim executable
This commit is providing a first version of an o2sim executable
which tries to integrate various modules together

We also provide a very basic SimConfig class for the purpose of
configuring or playing with the simulation from the command line.
This configuration is temporary and will be revised into a more
comprehensive scheme.