Adding table view abstraction for data sequences (#581)
* Adding table view abstraction for data sequences

Container class for multiple sequences of data wrapped by markers.

This is a container for data sequences of multiple frames consisting
of a header marker struct, a payload, and an optional trailer marker
struct. Each sequence forms a row in the TableView, the columns
are provided by the markers/frames.

A parser is used to step through the data sequence and extract
headers, trailers and payload positions.

Factoring out the test frame allocator utility to a separate file
to be used in multiple unit tests. Going to be made more general in
the future.

* Don't add the row desription for data sets without column structure