Resolve conflict of hidden device options with potential config specs.
The command line options are populated with a couple of hidden options
by the framework to correctly start the device. Using a key of the
hidden options in the processor config spec resulted in a parser error
due to duplicated definitions.

The allows in particular '--channel-config' to be used in the config
specs to specify external channels