Use LOG(INFO), rather than std::cout for commands
The current implementation of the driver has a naive way
of sending control requests from the child to the parent
via std::cout. It seems possible that sometimes this
is garbled by the actual Logger output, so I use the
Logger for the commands themselves assuming messages there
are atomic.