Allow driver state machine to perform callbacks
It is now possible to customise the behavior of the state machine via
callbacks. This allows, for example to express the dumping of DDS and
graphviz configuration files as state machine programs, rather than
ad-hoc functions.

This will also allow the GUI to be more flexible, e.g. to dump a text
representation of the current state, without having to introduce new
ad-hoc states.