Some prototype work on DPL input API, kept for record
NOTE: this approach has been abandoned in favor of solution in the following
commits, but we keep the prototype for the record.

Abandoned concept in this commit:
Apply a request-pointer-to-get-a-pointer policy for root-serializable objects.
The rule of thumb is now: to get an object with pointer-like behavior as return
one has to specify the pointer to type as template parameter.
- support extraction of messageable types with root dictionary by type, now
  a reference is returned to the (casted) payload content if the type is
  requested
- all serialized objects extracted by pointer
- also supporting containers of root-serializable objects by type in
  addition to pointer to type.

Additional check:
- checking serialization type when extracting messageable object by type
  to return a valid reference