Fix API issue with Lifetime
This fixes a serious API issue with Lifetime which could be erroneously
be used as subspecification due to the fact they both could decay to an
int. This makes sure that an explicit cast is required, hence raising
an error when a Lifetime is naively used in place of a subspecification.

This also armonises OutputSpec::Lifetime and InputSpec::Lifetime into a
single `o2::framework::Lifetime` since there is no reason for them to be
separate.

I've also cleaned up a few default places where the default
"Lifetime::Timeframe" was used to reduce verbosity of the examples /
tests.