SimReader device + TPC digitization using DPL
This commit provides a step/initial components towards
the processing of simulated hits using the O2 Data Processing Layer.

The main components are

a) SimReader:

A device which is supposed to read and preprocess hit files generated
by detector simulation. Main tasks are

  - reading hits from background + signal simulations
  - assigning/sampling collision times based on bunch-crossing parameters
  - sampling of background/signal scenarios
  - propagating generated information to consuming devices (such as digitization)

b) TPCDriftTimeDigitizer:

A device doing (sectorwise) digitization for the TPC.
The digitization is done in units of drift times which makes
the pileup handling convenient.

What this PR achieves:

  - Demonstration of a new parallel workflow using a realistic task
    from simulation

Note that these components are still under heavy development.
The features, names as well as file locations are still likely to change.