A flush method for DataAllocator
Extending the DataAllocator with a flush method.
This can be invoked by the user to explicitely ask for messages to be sent
now (instead implicitely at the end of a DataProcessor function).

With this function it should be possible to implement more generic data
forwarding patterns. At the moment a processing function is invoked on incoming
data and can only generate output data packet at the end of processing.
This might not be optimal for streaming/pipelining algorithms and for memory reasons.

We should now be able to forward data as soon as it is available.

A new test (flushTest) has been added, demonstrating the feature