Bugfix: set next-header flag in the last header of Stack as parameter
The argument list to built up a header stack can be an argument pack
of stacks and BaseHeader-derived headers. Since stacks consist of a
collection of headers, the next-header flag must be set in the last
header of the sequence. Until now, the flag was only set in the first
header (where it was already before).

Stack itsself is more generic and would also support e.g. arrays and
strings. However, in order to support header navigation, Stack is
restricted to contain only BaseHeader derived structures. Enforcing
this restriction now with a static_assert.