Introduce facility to debug the DataRelayer
As we develop the ability to customise the dispatching of computations
via the CompletionPolicy it became clear that a proper way of debugging
the state of the DataRelayer is fundamental to actually understand what
is going on with the system.

This introduces a simply GUI element which displays for each device the
state of the DataRelayer.