Improve log parsing speed for control and metrics
Regular expressions were avoided/optimized as much as possible. This code is
executed very frequently on the control process and profiling shows it had,
before optimization, a heavy impact on the workflow.

After optimization, the vast majority of the CPU time is spent in the regular
expression parsing metrics. There are still margins of optimization there.

Note that we rely on the fact that:

* control messages use log level "INFO"
* metrics use an unknown log level