DPL: Fix issue when merging multiple timeslices
A brown paperbag bug was introduced when moving from one channel per
product to one channel per device to device connection. The end result
was that whenever we were gathering timeslices from different time
parallel producer, the merger was actually considering phantom inputs
which were not really there but time-multiplexed. This fixes the issue
by doing the correct accounting for the datatypes we are waiting for.