Second step of the import of the MCH geometry (#1234)
* Complete the quadrant definitions, fix their placements and clean code

* Test if we find the same half-chambers position than AliRoot

* Change the quadrant creation for station 2

* Correct the volume assembly creation in the Station345 builder.

By changing the creation of the quadrant in the station 2 (see
previouscommit), we noticed that the "pile-up" of layers was incorrect.
This commit fixes the placement of the layers in the slat-stations

* Correct the frame volumes creation in station 2.

Some changes compared to AliRootby looking
atCERN-THESIS-2008-170(pages76-92). Some thingsare still
underinvestigation


* Change the definition of the air medium in the GeometryTest to fix an ID
conflict with the gas

* Don't print overlaps anymore

* Export the MCH geometry to a ROOT file

* Continue to import the station 1 geometry

* Correct the y position of the half-chambers.

Due to the inclination of the chambers with respect to the beam pipe, a
shift has to be applied. The y value is modified after investigation on
the computation done in AliRoot. The z shift is negligeable.

* Change the epoxy medium definition.

The chemical composition has been modified after looking at other
detector definitions. Moreover, the density has been decreased (to the
MFT one for the moment).

* Simplify the MANU positioning

* Put back the draw geometry function