Fixing sporadically failing unit test
We have to send at least one message to make sure that the callback of the consumer
is called to request termination via the ControlService. Depending on the scheduler,
the ControlService of the consumer might not be running when the producer sends the
termination command for all devices. And then termiation for the consumer is never
triggered.

This will however not fix the problem that sometimes the device does not terminate
correctly. The typical signature is that the device gets the terminate signal and
leaves the RUNNING state, entering READY state is the last message.