Fix trackID issue in FIT hits
TrackIDs for FIT hits where not correct in the
simulation file. We need to communicate to the stack
that a certain track left a hit, for this to be correct.

This commit fixes this for FIT, but we need to revise
all detectors. I guess this calls for a more framework-like
solution where the registration is done automatically.