DPL: saner rate for rate channel logger
Since real people are starting to use DPL, it turns out it is not
particularly user friendly, especially with many devices and without the
GUI, to have so many messages about the performance rate. As a quick fix
I hardcoded a saner value than the current one. This should probably be
made configurable in an easy way.