DPL: Initialize things in the init function and not elsewhere
Solving a memory problem in the digitizer workflow by moving the initialization
of the TPC digitizer into the actual init lambda.

Before it was done in the "getDataProcessorSpec" function. This is not a good place
since in fact every single DPL processor/process with evaluate this function (if it needs it
or not) upon topology creation.
In this case, every single DPL processor was instantiating pixels of a TPC readout chamber...
even though it was not actually doing TPC stuff.

We need to be careful about that. Some other data structures such as shared_ptrs need to be revised
in different places.

On the other hand, I think there is no need that every single processor evaluates the whole topology
at startup. This should really be done by a single master processor which communicates the topology
to the forks. This would also avoid such problems.