Adding termination policy and condition for RootTreeWriter
Adding an option to control the termination policy, available arguments are
'process' to terminate only the process, and 'workflow' to terminate all
processes. Default is 'process'
In fact, the processes are only marked as being 'ready' and the DPL driver
by default automatically terminates, if all processes are ready.

An option termination condition checker can be added. The checker consists of
a lambda/function working on DataRef objects.