Activating alternative lane configuration in order to use time slice concept
The TPC reconstruction workflow has so far mixed the concept of DPL time slices
with the 'geometrical' parallelsism. Using now the alternative lane configuration
which defines one channel per TPC sector and distributes those channels as multiple
inputs/outputs among the lane. With that, the time slice is preserved and the DPL
can do the correct matching of data sets. This allows to avoid buffering of inputs
in the tracker and also to apply time-parallelism on the tracker.