Improve CMake testing (O2-499)
Following discussions in WP3, tests are now wrapped by a script that handles
timeouts and retries. Retries are not handled by native CTest. It is possible to
ignore failures if some test has failed because of a timeout (not a CTest
feature either).

A separate logfile for each test is kept in the build directory. At the end of
each test, the logfile is renamed to include the exit code in the file name.
This makes it possible for the external process driving the build (alidist
recipe, aliBuild, the C.I.) to keep logfiles and make them available even in
case of success for further investigation. Timing out processes will have their
stack trace collected before they are killed.