Macro to build ITS cluster patterns dictionary
This macro uses full clusters tree to build their patterns dictionary. If the MC hit files is provided
it will be used to assign to patterns the mean bias between its COG and the MC truth, otherwise
the bias is set to 0 (as well as the cluster errors). Note that the clusters from the QED and noise
are contributing to patterns frequency calculation but do not contribute to bias estimates