Cleaning up input and output types of the TPC reco workflow
The data type 'clusters' has formerly been the cluster output of the digitizer.
Since the conversion step has been removed and the digitizer directly outputs
TPC ClusterHardware in raw pages, this output type has been removed but was still
in the help message of the option.

Data type 'decoded-clusters' is now renamed to 'clusters' and refers to the output
of the HardwareClusterDecoder, the TPC ClusterNative format.