Defining default move constructor and assignment operator for MCTruthContainer
Have been running into problems with some existing code because the content of
the container was not really moved to the target instance. By explicitly defining
the move constructor and assignment operator as default, the compiler does the
correct thing with the member vectors.

There was probably a bug in the TPC HardwareClusterDecoder where std::move(container)
has been used with the intention to make the original container empty.