Create timer message for timeslots which were created for that
Before the timer message was created in the first available slot
when the timer would get expired, instead of putting it in a slot
which was created for especially for that.
Fixing this bug exposed another issue - after some time, there were
many valid timeslots, eventually occupying all the cache. This resulted in
data messages not being not relayed, taken as obsolete. This will be fixed
 in another commit.