Fix desynchronizing of timer slot creation and timer expiration
The timer mechanism had to timers - one for booking a new timeslot,
another one for creating the message. After some time they would
desynchronize, which caused situations, when a timeslot was created
(by decision of creator callback), but the message itselft would not
be created (because expirator being a **little** bit late).

All of that resulted in having valid slots dedicated for timers, but
they were never filled with a timer message. After some time, all of
the slots were used by timers, and incoming data messages (with
enumerated timeslices) was interpreted as obsolete.