DPL: force run if --run option is provided
Because of the ways workflow merging works, if a dpl executable
is piped to a an external process, e.g. tee or less, it will
dump the configuration, rather than run. By passing the `--run`
option you change this behavior and force it to run regardless
of a pipe being present.