DPL: extract SoA based Analysis Data Model
This extracts a few of the SoA skins we have so far into their own
header file, so that it can be shared among different example analysis.

It also moves one of the analysis example (o2SimpleTrackAnalysis) to use
the SoA API (with rather obvious improvements in maintainability of the
analysis code itself).