[DPL] Updating WorkflowSpec::parallelPipeline helper
Updating the distribution algorithm of channels to the pipelines. The subspec
ids are now accessed round robin when distributed among the pipelines. This
will ensure a better performance distribution under the assumption that a
publishing process will publish the data sequentially. The data will thus be
distributed round robin among the pipelines. This will be naturally the
case unless the publisher itsself implements scheduling.