Getting rid of broadcast method for external FairMQ DPL proxies
The bradcast message has been used to forward the incoming messages from
an out-of-bound device on all relevant output channels. There was a bug,
as this can only work for the first channel, messages are unique pointers
which will be empty after first send.

Messages are now sent only on the concrete channel according to the OutputSpec
of the adapter function.

The relevant information is retreived via the RawDeviceContext and for this
purpose, the DeviceSpec has been added to RawDeviceContext.

The index of FairMQ channels has been removed as we use only one channel per
channel name.