DPL: optimize processDanglingInputs (#2538)
Before this processDanglingInputs does NxM complex pass on the
inputs and the expirators, checking if there is anything to be
expired and eventually expiring it.

This changes things so that we do Nxm complex pass, where m is
the actual inputs that can expire, i.e. we do not do the pass
for all the inputs which do not have an expiration policy attach.
Since not having an expiration policy is the norm, this should
minimize the impact of processDanglingInputs for normal usage.