Supporting also vector<messageable> in the RootTreeWriter
Implementing a specialization for vector<messageable> retrieving the input
opject as span and copying the data to the storage variable which is then
stored in the branch. Implementation for Vectors of non messageable types
with ROOT dictionary is unchanged.

TODO: Need to investigate if the binary branch specialization can be extended
to any POD pointer and realized without copy. Add read-back check for binary
branch test.