Printing error messages on output dispatch failure in ExternalFairMQDeviceProxy
The dispatch of output messages might be delayed or failing if downstream queue
is full or unconnected, using async sending in order to log error messages.
No messages are dropped at this stage, if the dispatch is not possible, also
upstream queues will be blocked. This is what we see in O2-962

TODO: Need to implement improved handling for failed send operations and dropped
messages. Maybe not at the level of DPL. This also applies to the DataProcessor
implementation, the heart of the DPL message dispatch. See O2-966.