Fix RawParser: set proper size estimate for payload
So far the fixed page size has been used to calculate the size of the payload,
simply because of the lack of information at the time of implmenting the parser.
The variable size of the payload is stored in the member memorySize in both
RDH v4 and v5. Since the raw parser does only supports those two versions at
the moment, the fix can be applied as it is. there is no need to support RDH v2
which has a member blockLength.