Simplifying the completion policy in TPC workflow
The custom completion policy was implemented in a complecated way to make
sure that data and mc of the same sector are available before starting the
computation. After moving to grouped output messages for one computation/each
triggered dispatch in commit e9c44429, we can rely on the fact that these
two inputs are arriving together.

The custom dispatch policy of the TPC reco workflow has been extended to trigger
dispatch on published digits/clusters/etc and publishing the optional mc data
before.

Minor correction in addition: fixing a compiler warning