DPL Util: RootTreeReader extension to support unserialized output
Branch configuration now allows specification of a type as template parameter
of the helper struct BranchDefinition. The type is used to deduce the optimal
serialization method. By default (without type), all objects are published
ROOT serialized. Messagable types with ROOT dictionary can now be published
unserialized if the type is provided in the reader configuration.