DPL raw input proxy: Skipping merging of split parts in the external proxy
This is a long pending task. Removing merging of split parts which had been
introduced in the first version of the raw proxy when multiple message per
route have not been supported by DPL.

Messages are now simply forwarded on the DPL route, independent of this it
needs to be checked what is/whether something is necessary to forward shm
messages without copy.