[EMCAL-645] Implemetation of an EMCAL event handler
Providing easy access to objects (clusters / cells /
cell indices) from the same trigger within a
timeframe. Can be initialized with cells, clusters
(both separated) and cluster + cells together,
together with the corresponding trigger records.
The event data contains either empty gsl::spans
for the branches not initialized or the gsl::spans
with the range of objects belonging to the same
event. The event handler is stl-iterable as
bidirectional iterator.