DPL: Using InputSpan as argument for CompletionPolicy callback
The change is transparent for the user code as the InputSpan provides an
iterator interface. Using InputSpan will allow to have access to the parts
within one route in order to create more complex completion checks.

Moreover, the extra copy of the cache column can be dropped, the entries of
the column are accessible through the span.