DPL RootTreeWriter: Adding support for runtime branch selection
Branch definitions need to be specified at compile time for setting up
all input types. Supporting now runtime selection to make the code simpler,
all definitions which might be used are specified at compile time and
the selections can be done on a runtime condition. Branch definitions
are disabled by setting the number-of-branches parameter to 0.