Extracting unserialized TPC digit data by gsl::span
The span is directly using the raw input data without copy, for writing an
std vector of digits the data needs to be copied to such an object.
TODO: we can investigate if it is possible with some custom schema evolution
code in the LinkDef to avoid this, but this is probably more of academic
interest.