[MCH] New definition of digit time and raw codec tests reorganisation
The new definition combines the SAMPA time stamp, the SAMPA bunch
crossing counter, and the orbit number of the HBFrame containing the
digit.

The new definition does not change the size of the Digit structure.

The introduction of this new time definition requires adding the bunch
crossing counter to the SampaCluster structure, and therefore to change
the Elink decoders as well as some of the unit tests.

Talking about unit tests, the tests based on reference buffers have been
removed.  The creation of those buffers was a bit too fragile and quite
cumbersome. Replaced with the closure codec test instead.