Extending the TPCSectorCompletionPolicy
Introducing Config flags with Config::RequireAll to require data from all
inputs in addition to the matching TPC sector policy. To be fully correct
we would need to require data from all inputs not going into the TPC policy,
but that is not possible for the moment. That's why there is a possibly
unhandled case if multiple TPC input routes are defined but a complete
data set is coming over one of them. Not likely to be a use case, though.

Usage:
  TPCSectorCompletionPolicy("processor-name-regexp",
                            TPCSectorCompletionPolicy::Config::RequireAll,
                            InputSpec{"", ConcreteDataTypeMatcher{"DET", "RAWDATA"}}, ...)();