Adjusting cluster output of the CA racker
Sending out the updated transport format which will allow to transport
the data between devices without copy. For the moment, the CA tracker
copies the data into the output message, but the GPU worker will be
changed to write directly to the provided buffer.

Further improvements to the input handling in the CA tracker spec:
- Get rid of the fixed arrays for collecting the input objects
- Adjusting cluster completeness check in the TPC CA tracker to use
  the sector bitmask from the sector header

The check can be dropped in the future as it is the task of the
completion policy to provide complete data sets. We keep it as a check
for a while during the hot development phase.