RawFileWriter allows intermediate caching of input data
If detector cannot ensure that the input data will be time ordered (e.g. TPC) then one can
activate the intermediate caching via TTree per link. In this case the actual superpages
filling will happen on calling RawFileWriter::Close and will ensure that the data of different
links going into the same file is ordered in time.
This is needed for data which will be used in readout.exe / DD replay, the RawFileReader can
read data with any ordering.