DPL: add service_ptr with policy based locking
Depending on the kind of service the user asks for, they will get
a smart pointer which includes a locking policy which will lock
(or not) the service on creation and unlock it on destruction. Because
services declare their (service) dependencies we are guaranteed that
two services will lock their dependencies in a compatible manner (otherwise
the topological sort of service dependencies will have loops).