RootTreeWriter: Zero copy treatment of trivially messageable vectors (#4417)
This commit provides a zero copy treatment for writing
std::vector messages to a TTree. Instead of copying the data once more,
the new approach fakes a std::vector appearance of the incoming
(shared) data (which was read as a gsl::span).

In result, this commit leads to memory improvements for
workflows such as digitization where the RootTreeWriter is used.