TPC: new const labels in digitizer workflow
Labels are now transported in shared memory as const flat buffer and written
to ROOT file without copy. With this we are now able to digitize a full
timeframe including labels.

The RootTreeReader is given a new hook to specialize publishing data.
This is applied to the conversion of flat labels to MCTruthContainer which
is still expected by the TPC reco workflow.

Next tasks will be the conversion of all read-only MCTruthContainer
containers to ConstMCTruthContainer in the remaining workflow.