o2-sim: More robust termination initiated by driver
So far, we've been relying on child-procs to return
correctly for termination. This has proven not to be
100% reliable (FairMQ sometimes hangs during shutdown).

The new treatment initiates the termination from the driver
via a callback-extension of the existing event-monitoring.

This actually provides a door-opener for more advanced
communication and triggering between all the parties involved.