Reduce memory footprint of TPC digitization in GRID mode
TPC digitization may require large RAM which may hit the limit of
16GB on 8-core GRID sites easily for moderate timeframe sizes.
Next to limiting the timeframe size, this poses also large constraints
on possible parallel scheduling with other tasks.

This commit introduces an **optional** alternative treatment of writing the data,
which will flush for any independent TPC drift-time chunk.

This IO flushing is done in the digitizer process itself, since it happens
at a smaller granularity than the DPL "run" invocation, and as such is
only useful for (GRID) productions in which we save intermediate
digit files.

The effect of this is the following:
a) Substantial reduction of memory, limited by the size of a drift-time.
   (We can stay on the order of 2GB even with 8 lanes). Here, TPC digitization
   may approach a 1-core "usability" limit.

b) Mem-reduction by a factor of 2 if we post-generate the full timeframe file
   in an after-burner.

c) Overall faster digitization/reconstruction because of faster ROOT IO.

To use the new feature one uses

i) `o2-sim-digitizer-workflow --onlyDet TPC --tpc-chunked-writer`
ii) `o2-tpc-chunkeddigit-merger | o2-tpc-reco-workflow --input-type digitizer ...`
(in which way we never do ROOT IO on large vectors)

instead of

i) `o2-sim-digitizer-workflow --onlyDet TPC`
ii) `o2-tpc-reco-workflow --input-type digits ...`