[EMCAL-610] Change handling of current bunch in bunch spitting
Changing find bunches method to push bunches
to the output vector after they are closed instead
of when they are opened. In this case buches with
0 length (ADCs suppressed by zero-suppression)
can be filtered out. A minimum amount of samples
(default: 3 - settable) is required to select the
bunch when closing.

In addition: Add documentation to header file.