TPC: Use OpenMP parallel digit publishing
Introducing parallelism into digit publishing,
for the case when digits were written in chunked mode.

Using OpenMP for this. Some critial sections protect
access to DPL ProcessingContext as well as when writing
to the shared mem container. (Maybe this could be avoided
by some mechanism.)

This significantly improves the startup phase of
TPC reconstruction on the GRID.