Check/enforce that all links got AddData call in every IR it processed
RawFileWriter will check for every IR for which at least 1 link was filled by detector
if all other registered links also got AddData call. If not, it will be enforced internally
with 0 payload and trigger and detField values provided in the 1st addData call of this IR.
This may be pose a problem in case detector needs to insert some header even for 0-size payloads.
In this case it is detector responsibility to make sure that all links receive their proper
addData call for every IR.