[MRRTF-117] Various Encoder fixes.
Those changes are intended to fix a number of issues in the MCH
digits-to-raw process. The changes are relatively numerous as some
refactoring was done.

The whole process sim->digits->raw->digits can be tested with the
`Detectors/MUON/MCH/Raw/test/codec-closure-test.sh` script. Note that a)
this script will be migrated to the full-system (extra) test somehow b)
it is _not_ currently passing because the digitizer has some issues
(that will be fixed in a forthcoming fix)