DPL: force processing of region callback without waiting for data (#6621)
Without this, it is possible that a late arriving event will not be processed
until some data actually arrives on one of the channels, unblocking the event
loop. This should obviate to the problem by triggering an aync libuv event
which will unblock the event loop without the need for data.