Implement --onlyDet option in o2-raw-tf-reader and enable reading of non-raw data
If detector run processing on the FLP, then the DD raw TF will contain derived data
(e.g. digits, triggers etc) instead of the raw.
For the detectors which might run some processing corresponding output channels are
defined (on top of the raw data channels).
These are TOF, a FIT detectors, EMC, PHS and CPV. For the EMC and PHS is it not yet
clear which subsecs will be sent from the FLP.