Introducing configurable channel name for the DPL output proxy (#7304)
The channel name can now be configured via workflow option '--proxy-channel-name'
of workflow o2-dpl-output-proxy. The new option defaults to 'downstream' which
has been the hardcoded name before.

There are probably more cases we need to look into as there are two proxy spec
generators in the ExternalFairMQDeviceProxy implementation, a consolidation is
planned.