Trace EMCAL run SOR/EOR in EMC DCS DP processor
At every run() invocation the method  checkRunStartEnd() will be called,
and the flag mRunStatus will be set to one of the following states:
EMCRunStatus::NONE       : there is no EMC run ongoing, mRunFollowed is at -1
EMCRunStatus::START      : the run with EMC just started, mRunFollowed contains the run number
EMCRunStatus::ONGOING    : the run with EMC continues, mRunFollowed contains the run number
EMCRunStatus::STOP       :  the run with EMC has just stopped,  mRunFollowed contains the run number.

The function will return the pointer on the GRPECSObject which is guaranteed to be non nullptr in
all cases except NONE. Note that the grp->getTimeEnd() will not have meaningful value until the
EMCRunStatus::STOP is encountered.

In the exceptional case when the workflow was started in the middle of the run with the EMCAL
(e.g. because it was restarted after the crash), you may see STOP without having START and
ONGOING before that (e.g. in case there is another run w/o EMCAL ongoing in parallel with the
EMCAL run: in this case you will see the just STOP when ECS will update the GRP of your run).

Note that the GLO/Config/GRPECS is still not populated by the ECS, so it cannot be tested.
For this reason I an option follow-emcal-run (default is false) is added.