Utility to adjust previous EOV to SOV of overriding CCDB object, use in Populator
If the CcdbObjectInfo of the object sent to the populator was created with
adjustableEOV flag (default), then, after uploading new object populator will
override the EOV of the previous overlapping object (if any) to the SOV of
the new object. This is done only provided the object to override has metadata
key \"adjustableEOV\" defined (value is irrelevant) and has no \"default\" key.
Newly uploaded object will have this \"adjustableEOV\" assigned in its metadata.