[EMCAL-784] Provide only EMV-EMCAL sampling factor for Geant4
Geant4 tune for EMCAL identical to the one used in AliDPG,
however the tag "EMV-EMCAL" was removed from the title.
Therefore the sampling factor "EMV" was selected as the
tag is part of the physics list and it is the second in
priorty. Choosing the EMV sampling factor leads to an
overscaling. In the current stage we only provide the
EMV-EMCAL scaling factor as it is adapted to the usual
physics list, in a next iteration we will provide dedicated
sampling factors for the other two physics lists as well.