MCTrack: change encoding for process to unsigned
We used 6bit int to be able to encode TMCProcesses inside MCTrack.
Unfortunately, when reading back as int this was interpreted in a signed
fashion and so an overflow/negative values were possible.

Now using an explicit unsigned 6 bit representation to fix the problem.
Fixes https://alice.its.cern.ch/jira/browse/O2-3030.