[EMCAL-630, EMCAL-789] EMCAL Reco Params and LGnoHG suppression
Use 3 sigma noise threshold with a sigma of 0.4 for
the LG digitizer to separate faulty cells from simple
noise: If the LG cell ADC is < 3 sigma the cell is
considered as pure noise cell where the error handling
is discarded, otherwise it is treated as a data corruption.

In addition moving converion from energy to ADC
from internal conversion constant to common constant. This has
a slightly better precision and will result in ~2% increase of
the energy.

[EMCAL-789] Create parameter handler for reconstruction params

Currently implemented:
- Time shift at cell level
- Noise cut LGnoHG