MCH: properly configure the slot finalization (#10566)
* [MCH] improve logic of slot finalization

The check of the slot statistics and the slot finalization are explicitly
done in the calibrator code instead of relying on the `checkSlotsToFinalize()`
function, because for MCH we only have one slot that spans the whole run.

`setUpdateAtTheEndOfRunOnly()` is also called at initialization to skip the
internal check of the slot statistics at each TF.

In the case where the calibrations are produced as soon as a target statistics
is reached, the subsequent data is skipped to avoid producing multiple
calibration objects for the same run.

* [MCH] use mapping in pedestals calibration

The electronics mapping is used to flag as valid the readout channels that
are associated to detector pads. Only valid channels are processed and
used to compute the total pedestals data size. Invalid channels therefore
will not contribute to the computation when checking the statistics of
a given calibration slot.

* [MCH] fix clang formatting errors

* MCH: update ped calib example script

* MCH: pedestal iterator only iterates on valid channels

* MCH: update testPedestalData test

Co-authored-by: Laurent Aphecetche <laurent.aphecetche@gmail.com>