Adapt tpc-track-study to real data, change output format
The output is split to 3 separate trees:
1) TTree tpcIni contains the initial track info:
a) counter     : global track counter (== entry of the track in this tree)
b) iniTrack    : original TrackTPC type track as it is in the reco.data, propagated to the reference radius (70 cm)
c) iniTrackRef : same track but refitted in the o2-tpc-track-study-workflow
d) time        : time0 of the iniTrack (in TPC TimeBins)
e) clSector, clRow, clX, clY, clZ : vectors with cluster data as it was evaluated for the iniTrack (original time0 assignment)

2) Tree tpcMC (present only if the workflow was running on MC data and w/o --disable-mc flag) with MC kinematics, contains:
a) counter     : global counter telling which entry in the tpcIni tree this track corresponds to. The counter is not guaranteed to be continuous: if the propagation of MC track fails, it will be omitted.
b) movTrackRef : reconstructed track (iniTrack) corresponding to this MC particle refitted with the MC true time assignment (imposedTB)
c) mcTrack     : corresponding MC particle converted to TrackPar format and propagated to the same radius as reconstructed tracks
d) imposedTB   : the MC track true timestamp in TPC TB used to refit the movTrackRef
e) dz          : naive change of the track Z under time0 -> imposedTB transformation, i.e. vDrift * (imposedTB - iniTrack.getTime0).
f) clX, clY, clZ : vectors with cluster data evaluated imposedTB time (sector and row data to be taken from the counter entry of the tpcIni)

3) TTree tpcMov: initial track evaluated for N (6) different time-stamps spanning allowed times of the track (tmin = iniTrack->getTime0() - iniTrack.getDeltaTBwd() : tmax = iniTrack->getTime0() + iniTrack.getDeltaTFwd())
a) counter    : global counter telling which entry in the tpcIni tree this track corresponds to. The counter is not guaranteed to be continuous: if the propagation of all tracks fails, it will be omitted.
b) copy       : id of the copy (spans 0 : maxCopy), i.e. evaluation is done for time = tmin + copy * (tmax - tmin) / maxCopy. Not guaranteed to be continuous since evaluation of some copies may fail.
c) maxCopy    : total number of timestamp considered for this track (uniformly covering tmin : tmax). Currently fixed to 5
d) movTrackRef: reconstructed track (iniTrack) refitted with the  time assignment (imposedTB) corresponding to copy above
e) imposedTB  : imposed time for this copy, see comment for 3b
f) dz         : naive change of the track Z under time0 -> imposedTB transformation, i.e. vDrift * (imposedTB - iniTrack.getTime0).
g) clX, clY, clZ : vectors with cluster data evaluated imposedTB time (sector and row data to be taken from the counter entry of the tpcIni)