DPL: introduce summary report support
The goal is to allow certain data processors, e.g. AOD reader
a callback which will be invoked at the end of the processing, in the
driver.

Such a callback will be provided with all the metrics which are
available and will produce some output summary for the whole run.