Distribution of CTP trigger within CollisionContext
This gives the (preliminary) possibility to distribute
CTPdigits (momentarily, trigger decisions for FT0 MVTX) for
digitizers.

This will work for MC that runs digitization in stages:
*) First FT0 + CTP
*) Then digitizers that trigger on FT0 MVTX

In a future PR, this can be done with proper DPL-inputs,
but this PR should already unblock work, where detectors need
access to CTP trigger records.

There is a small example in HMPID, showing how the CTP triggers
can be accessed.