DPL Analysis: avoid spurious message at the end of the processing
This solves the issue with the dummy extra iteration which is not
needed and actually flagged as an error after a8b3d4d221a.

The current logic is nevertheless to be reviewed because it seems
to silently skip all the timeframes in a file if the first table cannot be
read.