Expand MCTrack to contain proper energy property
Adding energy component to MCTrack properties.

Needs to be done since AOD also has energy
and we need to store generator particles for which
we don't know the mass (strings, resonances).

Despite adding an energy property to MCTrack,
we still use a hybrid mode:

a) for primaries, and few secondaries, we store the energy proper

b) for most secondaries, we store -1 in order
   and still use the mass from the TPDGDatabase to calculate the energy.
   This will ensure large compression and an almost zero impact on disc file size.
   We can do this since secondaries have a well defined mass.

In result, the energy of MC particles in AOD should now be completely trustable.