MCTrack: Generalize hit bit property
So far, every possible sensitive detector for ALICE Run3
had a dedicated bit assigned inside MCTracks. The number of such
bit was fixed to be 22 maximally.
In this way we could store information if a particular Monte Carlo track
caused energy deposits in detectors.

Since more and more detectors are being added to the code
(ALICE3, ITS3, R&D), we'll soon have more than 22 DetIDs in our code
which will conflict with the space available in MCTrack.

Since enlarging the MCTrack class in memory is to be avoided, this commit
proposed a solution based on intermediate lookup tables which dynamically
maps the detector ID to a particular bit.

This solution should scale much further than before. The limit is now 22 simultaneous
sensitive detectors in the geometry and not 22 total sensitive detectors in the code.