MatLUT: faster phi-sector determination; reduction of divisions
Following ticket https://alice.its.cern.ch/jira/browse/O2-4170,

Profile analysis has shown that material budget calculations used in tracking, spent most time in
(a) determination of a cylindrical layer
(b) determination of a phi sector in a layer

This commit proposes an improvement for (b):

- Instead of using system function phi = atan2(y, x), we use a
  faster - approximate version. We can do this, since we never need
  to use the precise result of phi in a calculation. Instead, phi is merely used
  to lookup a slice/sector.
  The error of the approximate version is ~1e-5, which is enough to get the correct phi
  sector in 99.999 of the cases.

  (Rare cases, where we end up in a wrong neighbouring phi sector are probably ok ... since
   the material budget lookup is anyways an approximation.)

In addition, we propose a smaller code mofication for ray.CrossZ to reduce
the number of divisions and if statements.
A previous check for
ray orthogonal to z can not happen in the code since there is a previous condition
`if (zID != zIDLast)`.