AOD: Fix segfault when MC kinematics event is empty
Fixing a segfault problem, arising from the fact
that the kinematics events is empty.

AOD conversion now correctly handles this cornercase.

mToStore is refactored to **not use** pointers to maps
so that we do not need to treat nullptr corner cases.

P.S. Simulation should probably make sure
that empty events are handled at an earlier stage.