AOD: Prepare Thinner for TPC-only V0s (#12510)
This patch carefully checks which tracks are associated with a V0 and flags to
not be discarded. This is written such that the thinner is unaware of the track-type.
Additionally, this patch further generalizes the thinner by making it check
manually for the table_name with regex. E.g., previously the v0-name was
specifically ‘O2v0_001’, now this has been changed to ’O2v0_00?’. That way, we
no longer care about the actual version contained in the table. Hence, new AO2Ds
can also be thinned.

Signed-off-by: Felix Schlepper <felix.schlepper@cern.ch>