Force digitisation when not in GRP
The reco workflow in data taking is agnostic, meaning that empty reco
files are written for detectors that were not readout in data taking.

To be able to resemble this in simulation, it should be possible to
always run the digitisation workflow for desired detectors
(--onlyDet, --skipDet) even if they are not in the GRP file which has
previously been created by the transport code.

This will open up the possibility to implicitly modularies our
simulation workflow in the following sense:

Digits will be present for all detectors that produced a hit file.
On the other hand, empty digi files will be produced for ALL detectors
such that the reco->match->AOD files will end up with the correct
content derived from detectors that actually produced hit files.

This will drastically reduce the need of specialising our simulation
workflow graphs.
In particular, this is a step forward to more resilient anchored
simulation.