Only override LD_LIBRARY_PATH for rootcling, not for CMake (#12683)
CMake depends on more things that are in the customised LD_LIBRARY_PATH, such
as OpenSSL. Therefore, the new LD_LIBRARY_PATH must only apply to the
rootcling command, not to CMake.

Assigning to an exported variable in bash changes the exported definition, so
assigning to LD_LIBRARY_PATH directly (even without explicitly exporting)
means the variable already takes effect for the wrapping cmake command. This
does not seem to be intended, so use a temporary variable instead.