DPL: allow Signposts to work as LOGF(level) replacement
This introduces a specialised API to emit signposts which can
be used in place of LOG(info), LOG(warning) , LOG(error).

Using it, errors and warning messages will appear as signposts as well (matching
the correct interval and so on), while the info messages will appear as signposts
if the signpost is enabled, without duplication, simplifying grepping.