Fix Z bias of TRD-matched tracks
1) When the time bias of TPC timebins was introduced, it had to be explicitly accounted for in the TPC tracks refits constrained by TRD timestamps (the t0 was subtracted from the TRD time).
When the TPC fitter started to account for this t0 bias provided via VDrift object, this explicit accounting had to be removed but it was not.
As a result, the t0 bias was accounted for twice, leading to a few cm bias in Z for constrained TPC tracks.
While we were using large (1cm) systematic errors on Z error of TPC tracks, this bias was overridden by the following refit in the TRD or in the ITS,
once we adopted smaller Z-error the bias started to affect matching in the TOF.

2) The TPC refitter was called on the track at the outer ITS or inner TRD radius (for the outward and inward refits resp.). Since the refitter propagates w/o material corrections,
effectively the tracks were not accounting for the material between the TPC and ITS/TRD (before the refit, so the effect was small anyway).
Now the tracks are explicitly propagated befor the refit to the TPC boundary using the standard propagator accounting for materials.