o2-sim: Ability to focus/select a particular chunk to transport
For debugging purposes, it is sometimes of advantage to skip
to a particular event chunk for debugging.
This commit allows to select an event-ID + event-Part for transport.
Everything else will be skipped. This is useful in situations where
we know of a bug in certain event and we'd would like to study this-and-only-this
event (under same seeding).

The event-part selection is done via env variable `O2SIM_RESTRICT_EVENTPART=event:part`