Fix for HepMC: Actually take into account mEventsToSkip
The option HepMC.eventsToSkip existed with the purpose
to skip the first couple of events. This is useful
to share an HepMC file among multiple timeframes or MC-jobs.
Jobs can skip ahead to the correct event.

Unfortunately, the skipping itself was not implemented
and the option had no effect. This commit is fixing this.

In addition, this commit avoids a potential infinite hang
under error conditions.