CCDB: Ensure CCDBDownloader's libuv loop is properly deleted (#13238)
My valgrind checks were showing that the internal CCDBDownloader's libuv loop resources were not fully released. With this fix, valgrind is happy.

The change is based on the following premises:
- `uv_loop_alive` returns true if there are *active* handles
- `uv_loop_close` returns `UV_EBUSY` if there are *open* handles.
- we need to succesfully call `uv_loop_close` to clear `mUVLoop` resources

Thus, in case that we would not have *active* handles, we would never close *open* ones, and consequently would not manage to call `uv_loop_close` without getting `UV_EBUSY`.