DPL Analyis: introduce value based has_type
Using the std::disjunction is recursive in the pack elements,
therefore slow. This version allow to expand the pack without recursive
type creations and seems to be compiled much faster.

We keep the old version around because it's still needed in some places
in O2Physics and therefore we need to do an incremental migration.