feat(CCDB): add getForRunMiddleTimeStamp method
Adds a new method `getForRunMiddleTimeStamp` to the `CCDBManagerInstance`
class. This method retrieves an object of type `T` from the CCDB using the
path and the middle timestamp of the given run number. This is a useful
utility method for cases where the exact timestamp does not matter and the
run number is available.
Optionally use metadata in the query
Rename the function
Fix setting of the metadata