MCH: introduce digit modifier in filtering workflow (#13924)
* [MCH] introduce digit modifier in filtering workflow

The digit modifier allows to change the contents
of the digits in the filtering step.

It is introduced in order to correct some mapping
issues in the CTFs already collected, but the
interface is general, and in the future it might be
used for any kind of digit manipulation, if needed.

* [MCH] added pad remapping function for ST1

The remapping function corrects the pads mapping
for seven DS boards on the edge of the bending
cathodes, all corresponding to the motif type "1BG".

* [MCH] added pad remapping function for ST2

The remapping function corrects the pads mapping
for five consecutive motif types on the bending
side of ST2 quadrants (types "2Bv1" to "2Bv5").