HepMC: Remove default backgrounding mode
Removing the default & at the end of HepMC child processes.
This should not be needed and actually leads to children be executed
outside of the process tree of simulation. This in turn corrupts
CPU and MEM monitoring in O2DPG MC workflows.

Changing default for now but maybe we can eventually get rid of this completely.