Make sure extra tracks are randomized to avoid PHOS hole losses
The tracks with TRD/TOF are prioritized as of higher quality and we accept a limited min. amount of tracks per TF (1)
before swithing to extra tracks (whose max accepted number is also capped).
Therefore this guaranteed minimum will be typically affected by the PHOS hole, the same is true for the ITS-TPC stripped
versions of these global tracks which are added for additional processing if ITS only extrapolation is asked.

When processing extra tracks, the algorithm was considering only these stripped ITS-TPC tracks,
so the ITS extrapolation were also affected by the PHOS hole.
This PR makes sure that the extra tracks processing starts from the first seed abandoned due to meeting the condition (1),
and that the remaining tracks are processed in the random order.