Glo: only initialzed AB geom helper up to lowest layer
We can skip initializing the layers below the lowest allowed AB layer.
This allows to run the AB for ITS3 otherwise the initialization of the
RecoGeomHelper is not well defined and crashed for layers < 3 since the
chip mapping is not well defined.
Currently we anyways allow AB tracks only to go down to layer 3 by
default and this might not change for Run 4 anytime soon. If it does the
RecoGeomHelper class has to be adapted then.

Signed-off-by: Felix Schlepper <felix.schlepper@cern.ch>